Moved @tryghost/link-tracking to core
fix https://linear.app/ghost/issue/ENG-2064/link-tracking

- this commit moves the code and tests for @tryghost/link-tracking into
  Ghost core so we can cut down on the code indirection and keep code
  together
- moves files, fixes paths and removes references
