+++ compose.yml
@@ -71,7 +71,6 @@ services:
       - node_modules_ghost_job-manager:/home/ghost/ghost/job-manager/node_modules:delegated
       - node_modules_ghost_link-redirects:/home/ghost/ghost/link-redirects/node_modules:delegated
       - node_modules_ghost_link-replacer:/home/ghost/ghost/link-replacer/node_modules:delegated
-      - node_modules_ghost_link-tracking:/home/ghost/ghost/link-tracking/node_modules:delegated
       - node_modules_ghost_magic-link:/home/ghost/ghost/magic-link/node_modules:delegated
       - node_modules_ghost_mail-events:/home/ghost/ghost/mail-events/node_modules:delegated
       - node_modules_ghost_mailgun-client:/home/ghost/ghost/mailgun-client/node_modules:delegated
@@ -246,7 +245,6 @@ volumes:
   node_modules_ghost_job-manager: {}
   node_modules_ghost_link-redirects: {}
   node_modules_ghost_link-replacer: {}
-  node_modules_ghost_link-tracking: {}
   node_modules_ghost_magic-link: {}
   node_modules_ghost_mail-events: {}
   node_modules_ghost_mailgun-client: {}
