+++ ghost/core/package.json
@@ -115,7 +115,6 @@
     "@tryghost/limit-service": "1.2.14",
     "@tryghost/link-redirects": "0.0.0",
     "@tryghost/link-replacer": "0.0.0",
-    "@tryghost/link-tracking": "0.0.0",
     "@tryghost/logging": "2.4.21",
     "@tryghost/magic-link": "0.0.0",
     "@tryghost/mail-events": "0.0.0",
