--- ghost/link-tracking/lib/link-tracking.js
@@ -1,6 +1,0 @@
-module.exports = {
-    LinkClickTrackingService: require('./LinkClickTrackingService'),
-    LinkClick: require('./ClickEvent'),
-    PostLink: require('./PostLink'),
-    FullPostLink: require('./FullPostLink')
-};
